Make DraftYear and SeasonYear sealed; update nullable strings
The DraftYear and SeasonYear classes have been changed to sealed
classes, preventing inheritance. This is indicated by the addition
of the sealed keyword in DraftYear.cs and SeasonYears.cs.

In PlayerDraftRanking.cs, several properties in the PlayerDraftRanking
class have been updated to allow nullable strings. This is indicated
by the addition of the ? after the string type for properties such as
LastName, FirstName, PositionCode, ShootsCatches, LastAmateurClub,
LastAmateurLeague, BirthDate, BirthCity, BirthStateProvince, and
BirthCountry.